+++ src/reply/send_reply.rs
@@ -27,7 +27,7 @@ pub async fn send_reply<'att, U, E>(
     builder: impl for<'a> FnOnce(&'a mut crate::CreateReply<'att>) -> &'a mut crate::CreateReply<'att>,
 ) -> Result<crate::ReplyHandle<'_>, serenity::Error> {
     Ok(match ctx {
-        crate::Context::Prefix(ctx) => crate::ReplyHandle(super::ReplyHandleInner::Prefix(
+        crate::Context::Prefix(ctx) => super::ReplyHandle(super::ReplyHandleInner::Prefix(
             crate::send_prefix_reply(ctx, builder).await?,
         )),
         crate::Context::Application(ctx) => crate::send_application_reply(ctx, builder).await?,
@@ -54,30 +54,20 @@ pub async fn send_application_reply<'att, U, E>(
     ctx: crate::ApplicationContext<'_, U, E>,
     builder: impl for<'a> FnOnce(&'a mut crate::CreateReply<'att>) -> &'a mut crate::CreateReply<'att>,
 ) -> Result<crate::ReplyHandle<'_>, serenity::Error> {
-    let mut data = crate::CreateReply {
-        ephemeral: ctx.command.ephemeral,
-        allowed_mentions: ctx.framework.options().allowed_mentions.clone(),
-        ..Default::default()
-    };
-    builder(&mut data);
-    _send_application_reply(ctx, data).await
+    _send_application_reply(ctx, ctx.reply_builder(builder)).await
 }
 
 /// private version of [`send_application_reply`] that isn't generic over the builder to minimize monomorphization-related codegen bloat
 async fn _send_application_reply<'a, U, E>(
     ctx: crate::ApplicationContext<'a, U, E>,
-    mut data: crate::CreateReply<'_>,
+    data: crate::CreateReply<'_>,
 ) -> Result<crate::ReplyHandle<'a>, serenity::Error> {
     let interaction = match ctx.interaction {
         crate::ApplicationCommandOrAutocompleteInteraction::ApplicationCommand(x) => x,
         crate::ApplicationCommandOrAutocompleteInteraction::Autocomplete(_) => {
-            return Ok(crate::ReplyHandle(super::ReplyHandleInner::Autocomplete))
+            return Ok(super::ReplyHandle(super::ReplyHandleInner::Autocomplete))
         }
     };
-
-    if let Some(callback) = ctx.framework.options().reply_callback {
-        callback(ctx.into(), &mut data);
-    }
 
     let has_sent_initial_response = ctx
         .has_sent_initial_response
@@ -108,7 +98,7 @@ async fn _send_application_reply<'a, U, E>(
         None
     };
 
-    Ok(crate::ReplyHandle(crate::ReplyHandleInner::Application {
+    Ok(super::ReplyHandle(super::ReplyHandleInner::Application {
         http: &ctx.serenity_context.http,
         interaction,
         followup,
@@ -120,24 +110,14 @@ pub async fn send_prefix_reply<'att, U, E>(
     ctx: crate::PrefixContext<'_, U, E>,
     builder: impl for<'a> FnOnce(&'a mut crate::CreateReply<'att>) -> &'a mut crate::CreateReply<'att>,
 ) -> Result<Box<serenity::Message>, serenity::Error> {
-    let mut reply = crate::CreateReply {
-        ephemeral: ctx.command.ephemeral,
-        allowed_mentions: ctx.framework.options().allowed_mentions.clone(),
-        ..Default::default()
-    };
-    builder(&mut reply);
-    _send_prefix_reply(ctx, reply).await
+    _send_prefix_reply(ctx, ctx.reply_builder(builder)).await
 }
 
 /// private version of [`send_prefix_reply`] that isn't generic over the builder to minimize monomorphization-related codegen bloat
 async fn _send_prefix_reply<'a, U, E>(
     ctx: crate::PrefixContext<'_, U, E>,
-    mut reply: crate::CreateReply<'a>,
+    reply: crate::CreateReply<'a>,
 ) -> Result<Box<serenity::Message>, serenity::Error> {
-    if let Some(callback) = ctx.framework.options().reply_callback {
-        callback(ctx.into(), &mut reply);
-    }
-
     // This must only return None when we _actually_ want to reuse the existing response! There are
     // no checks later
     let lock_edit_tracker = || {