+++ src/structs/context.rs
@@ -161,17 +161,6 @@ context_methods! {
         self.serenity_context()
     }
 
-    // Poise internals can't pass `poise::Context` into `impl CacheHttp` parameters, because
-    // `CacheHttp: Send + Sync` so poise internals would need to add noisy `U: Send + Sync`
-    // everywhere.
-    // Poise internals also can't continue to call `.discord()` because it's deprecated. And
-    // `.serenity_context()` is too long. So we have this instead.
-    #[doc(hidden)]
-    (sc self)
-    (pub fn sc(self) -> &'a serenity::Context) {
-        self.serenity_context()
-    }
-
     /// Returns a view into data stored by the framework, like configuration
     (framework self)
     (pub fn framework(self) -> crate::FrameworkContext<'a, U, E>) {
@@ -251,7 +240,7 @@ context_methods! {
             ctx.interaction.member().map(Cow::Borrowed)
         } else {
             self.guild_id()?
-                .member(self.sc(), self.author().id)
+                .member(self.serenity_context(), self.author().id)
                 .await
                 .ok()
                 .map(Cow::Owned)
@@ -439,6 +428,45 @@ context_methods! {
             Context::Application(ctx) => Some(ctx.interaction.locale()),
             Context::Prefix(_) => None,
         }
+    }
+
+    /// Builds a [`crate::CreateReply`] by combining the builder closure with the defaults that were
+    /// pre-configured in poise.
+    ///
+    /// This is primarily an internal function and only exposed for people who want to manually
+    /// convert [`crate::CreateReply`] instances into Discord requests.
+    (reply_builder self builder)
+    (pub fn reply_builder<'att>(
+        self,
+        builder: impl for<'b> FnOnce(&'b mut crate::CreateReply<'att>) -> &'b mut crate::CreateReply<'att>,
+    ) -> crate::CreateReply<'att>) {
+        let mut reply = crate::CreateReply {
+            ephemeral: self.command().ephemeral,
+            allowed_mentions: self.framework().options().allowed_mentions.clone(),
+            ..Default::default()
+        };
+        builder(&mut reply);
+        if let Some(callback) = self.framework().options().reply_callback {
+            callback(self, &mut reply);
+        }
+        reply
+    }
+
+    /// Returns serenity's cache which stores various useful data received from the gateway
+    ///
+    /// Shorthand for [`.serenity_context().cache`](serenity::Context::cache)
+    #[cfg(feature = "cache")]
+    (cache self)
+    (pub fn cache(self) -> &'a serenity::Cache) {
+        &self.serenity_context().cache
+    }
+
+    /// Returns serenity's raw Discord API client to make raw API requests, if needed.
+    ///
+    /// Shorthand for [`.serenity_context().http`](serenity::Context::http)
+    (http self)
+    (pub fn http(self) -> &'a serenity::Http) {
+        &self.serenity_context().http
     }
 }