+++ CHANGELOG.md
@@ -1,3 +1,36 @@
+# 0.5.5
+
+New features:
+- Added `#[min_length]` and `#[max_length]` support for slash command string parameters
+
+Detailed changelog: https://github.com/kangalioo/poise/compare/v0.5.4...v0.5.5
+
+# 0.5.4
+
+API updates:
+- The `payload` field of `FrameworkError::CommandPanic` has been changed from `Box<dyn Any + Send>` to `Option<String>`
+  - This is technically a breaking change
+  - However, the newly introduced `payload` field in 0.5.3 made `FrameworkError` accidentally not Sync anymore
+  - And `FrameworkError::CommandPanic` has only been introduced a few days ago in 0.5.3
+  - Therefore, I think it's ok to release this as a patch release to reverse the accidental breaking change from 0.5.3
+
+Detailed changelog: https://github.com/kangalioo/poise/compare/v0.5.3...v0.5.4
+
+# 0.5.3
+
+New features:
+- Added `builtins::paginate()` as an example implementation of pagination
+- Added missing events in `EventWrapper` (#144)
+- Added `FrameworkError::CommandPanic` to allow custom handling of panics (#140)
+  - `builtins::on_error` responds with an "Internal error" embed when encountering `CommandPanic`
+
+Behavior changes:
+- `builtins::on_error` now prints `FrameworkError::Command` not just in Discord chat, but in console as well
+  - because responding in Discord sometimes doesn't work, see 0a03fb905ca0bc3b2ee0701fe35d3c89ecf5a654
+- Fixed a compile error when `name_localized` or `description_localized` are used multiple times (#143)
+
+Detailed changelog: https://github.com/kangalioo/poise/compare/v0.5.2...v0.5.3
+
 # 0.5.2
 
 New features: